requirements/base.txt
django-extensions==2.0.6
djangorestframework==3.12.2
django==3.1.3
celery==5.0.2
ipython==5.6.0
django-utils-six==2.0
django-configurations==2.2
psycopg2-binary==2.8.6
django-redis==4.12.1
google-api-python-client==1.12.8
